Rework test report logic
- Use only INTEGER data, for simplicity
- Add shell script to generate report
- Add jq filters to aid in report generation
- Store test data in json file, not in db

+++ exercises/concept/intro-select/create_fixture.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS weather_readings;
 CREATE TABLE weather_readings (
   date        TEXT    NOT NULL,
   location    TEXT    NOT NULL,
-  temperature REAL    NOT NULL,
+  temperature INTEGER NOT NULL,
   humidity    INTEGER NOT NULL
 );
